timing testing vsync- off compatible code.
FossilOrigin-Name: 81acbb1a8c1110a443421b375b472c2d64fd6164a039f2b3b6e9666a9d337664

+++ frameserver/arcan_frameserver.h
@@ -28,7 +28,7 @@ extern FILE* logdev;
 
 /* to get rid of a few POSIX calls in libretro implementation */
 void* frameserver_getrawfile(const char* resource, ssize_t* ressize);
-unsigned long int frameserver_timemillis();
+unsigned long long int frameserver_timemillis();
 void frameserver_delay(unsigned long);
 
 #endif
